made adjustments from code review

+++ src/main/java/com/hedera/services/exchange/Rate.java
@@ -3,13 +3,13 @@ package com.hedera.services.exchange;
 public class Rate {
     private final Integer hbarEquiv;
     private final Double centEquiv;
-    private final Double expirationTime;
+    private final long expirationTime;
 
     public Rate() {
-        this(0, 0.0, 0.0);
+        this(0, 0.0, 0);
     }
 
-    public Rate(Integer hbarEquiv, Double centEquiv, Double expirationTime){
+    public Rate(Integer hbarEquiv, Double centEquiv, long expirationTime){
         this.hbarEquiv = hbarEquiv;
         this.centEquiv = centEquiv;
         this.expirationTime = expirationTime;
